move LoadSettings to models/settings.go
